// types/supabase.types.ts
export type Database = {
// ...
    public: {
        Tables: {
// ...
            recipe_ingredients: {
                Row: {
                    created_at: string | null
                    id: string
                    ingredient_id: string
                    quantity: number
                    recipe_id: string
                }
                Insert: {
                    created_at?: string | null
                    id?: string
                    ingredient_id: string
                    quantity: number
                    recipe_id: string
                }
                Update: {
                    created_at?: string | null
                    id?: string
                    ingredient_id?: string
                    quantity?: number
                    recipe_id?: string
                }
                Relationships: [
                    {
                        foreignKeyName: "recipe_ingredients_ingredient_id_fkey"
                        columns: ["ingredient_id"]
                        isOneToOne: false
                        referencedRelation: "ingredients"
                        referencedColumns: ["id"]
                    },
                    {
                        foreignKeyName: "recipe_ingredients_recipe_id_fkey"
                        columns: ["recipe_id"]
                        isOneToOne: false
                        referencedRelation: "recipes"
                        referencedColumns: ["id"]
                    },
                ]
            }
            recipes: {
                Row: {
                    created_at: string | null
                    dish_id: string
                    id: string
                }
                Insert: {
                    created_at?: string | null
                    dish_id: string
                    id?: string
                }
                Update: {
                    created_at?: string | null
                    dish_id?: string
                    id?: string
                }
                Relationships: [
                    {
                        foreignKeyName: "recipes_dish_id_fkey"
                        columns: ["dish_id"]
                        isOneToOne: true
                        referencedRelation: "dishes"
                        referencedColumns: ["id"]
                    },
                ]
            }
// ...
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            [_ in never]: never
        }
        CompositeTypes: {
            [_ in never]: never
        }
    }
}